ci(quality): split Bolt intercepts into backend/intercepts.rs
backend.rs crossed the 2500-line file ceiling when the plan
string-representation renderer landed. The intercept machinery — verb
recognizers, column contracts, stream builders, and the EXPLAIN-rows →
summary.plan conversion (~350 lines of free functions and consts) —
moves to backend/intercepts.rs; the Backend impl keeps the handlers
that need server state. No behavior change; bolt suite green.
